Archives/requirements.txt
duckduckgo-search
crawl4ai
openai
sqlmodel
python-dotenv